+++ src/core/widget_store.rs
@@ -59,24 +59,6 @@ impl WidgetStore {
         }
     }
 
-    /// Handles the resizing of the texture buffer after the window resize has taken place.  The
-    /// behavior should be processed before drawing is rendered, so the sequence of events should
-    /// be `event` -> `handle_resize` -> `invalidate` -> `draw`.  This is mainly handled by the
-    /// `pushrod::core::main` loop, but it can be handled programmatically if required.
-    pub fn handle_resize(&mut self, width: u32, height: u32) {
-        self.widgets[0].widget.set_size(width as i32, height as i32);
-
-        for pos in 0..self.widgets.len() {
-            self.window_resized_for_id(
-                pos as i32,
-                crate::core::point::Size {
-                    w: width as i32,
-                    h: height as i32,
-                },
-            );
-        }
-    }
-
     /// Invalidates all widgets in the window.  This is used to force a complete refresh of the
     /// window's contents, usually based on a timer expiration, or a window resize.  Use with
     /// care, as this is an expensive operation.
@@ -215,70 +197,6 @@ impl WidgetStore {
                 self.draw(paint_id, c, g);
             }
         }
-    }
-
-    /// Callback to `key_pressed` for a `Widget` by ID with its corresponding key, and button
-    /// state (pressed or released)
-    pub fn keypress_for_id(&mut self, id: i32, key: &Key, state: &ButtonState) {
-        &self.widgets[id as usize].widget.key_pressed(id, key, state);
-    }
-
-    /// Callback to `mouse_entered` for a `Widget` by ID.
-    pub fn mouse_entered_for_id(&mut self, id: i32) {
-        &self.widgets[id as usize].widget.mouse_entered(id);
-    }
-
-    /// Callback to `mouse_exited` for a `Widget` by ID.
-    pub fn mouse_exited_for_id(&mut self, id: i32) {
-        &self.widgets[id as usize].widget.mouse_exited(id);
-    }
-
-    /// Callback to `mouse_scrolled` for a `Widget` by ID, with the mouse scroll `Point`.
-    pub fn mouse_scrolled_for_id(&mut self, id: i32, point: Point) {
-        &self.widgets[id as usize].widget.mouse_scrolled(id, point);
-    }
-
-    /// Callback to `mouse_moved` for a `Widget` by ID, with the mouse position at `Point`.  The
-    /// mouse point is relative to the `Widget` itself, not its position on the screen.
-    pub fn mouse_moved_for_id(&mut self, id: i32, point: Point) {
-        let origin = &self.widgets[id as usize].widget.get_origin();
-        let new_point = Point {
-            x: point.x - origin.x,
-            y: point.y - origin.y,
-        };
-
-        &self.widgets[id as usize].widget.mouse_moved(id, new_point);
-    }
-
-    /// Callback to `window_resized` for a `Widget` by ID, with the new `Size` of the window.
-    pub fn window_resized_for_id(&mut self, id: i32, size: crate::core::point::Size) {
-        &self.widgets[id as usize].widget.window_resized(id, size);
-    }
-
-    /// Callback to `focused` for a `Widget` by ID.
-    pub fn handle_focus(&mut self, focus: bool) {
-        for id in 0..self.widgets.len() as i32 {
-            &self.widgets[id as usize].widget.window_focused(id, focus);
-        }
-    }
-
-    /// Callback to `button_down` for a `Widget` by ID, with the button code.
-    pub fn button_down(&mut self, id: i32, button: Button) {
-        &self.widgets[id as usize].widget.button_down(id, button);
-    }
-
-    /// Callback to `button_up_inside` for a `Widget` by ID, with the button code.
-    pub fn button_up_inside(&mut self, id: i32, button: Button) {
-        &self.widgets[id as usize]
-            .widget
-            .button_up_inside(id, button);
-    }
-
-    /// Callback to `button_up_outside` for a `Widget` by ID, with the button code.
-    pub fn button_up_outside(&mut self, id: i32, button: Button) {
-        &self.widgets[id as usize]
-            .widget
-            .button_up_outside(id, button);
     }
 
     /// Retrieves a reference to the `Box`ed `Widget` object by its ID.