+++ src/widget/image_widget.rs
@@ -15,7 +15,6 @@
 
 use piston_window::*;
 
-use crate::core::callbacks::*;
 use crate::widget::config::*;
 use crate::widget::widget::*;
 
@@ -47,7 +46,6 @@ use crate::widget::widget::*;
 /// ```
 pub struct ImageWidget {
     config: Configurable,
-    callbacks: CallbackStore,
     image: Box<G2dTexture>,
     image_size: crate::core::point::Size,
 }
@@ -72,7 +70,6 @@ impl ImageWidget {
 
         Self {
             config: Configurable::new(),
-            callbacks: CallbackStore::new(),
             image: Box::new(texture.clone()),
             image_size: crate::core::point::Size {
                 w: texture.clone().get_size().0 as i32,
@@ -87,10 +84,6 @@ impl ImageWidget {
 impl Widget for ImageWidget {
     fn config(&mut self) -> &mut Configurable {
         &mut self.config
-    }
-
-    fn callbacks(&mut self) -> &mut CallbackStore {
-        &mut self.callbacks
     }
 
     /// Draws the contents of the widget.