+++ src/widget/toggle_button_widget.rs
@@ -15,7 +15,6 @@
 
 use piston_window::*;
 
-use crate::core::callbacks::*;
 use crate::core::point::*;
 use crate::widget::box_widget::*;
 use crate::widget::config::*;
@@ -64,7 +63,6 @@ pub type MutableSelectedCallback = Box<FnMut(bool) -> ()>;
 /// ```
 pub struct ToggleButtonWidget {
     config: Configurable,
-    callbacks: CallbackStore,
     base_widget: BoxWidget,
     text_widget: TextWidget,
     selected_state: bool,
@@ -82,7 +80,6 @@ impl ToggleButtonWidget {
     ) -> Self {
         Self {
             config: Configurable::new(),
-            callbacks: CallbackStore::new(),
             base_widget: BoxWidget::new(),
             text_widget: TextWidget::new(
                 factory,
@@ -173,10 +170,6 @@ impl Widget for ToggleButtonWidget {
         &mut self.config
     }
 
-    fn callbacks(&mut self) -> &mut CallbackStore {
-        &mut self.callbacks
-    }
-
     /// Sets the `Point` of origin for this widget and the base widget, given the X and Y
     /// coordinates.  Invalidates the widget afterward.
     fn set_origin(&mut self, x: i32, y: i32) {
@@ -214,61 +207,61 @@ impl Widget for ToggleButtonWidget {
         self.base_widget.get_color()
     }
 
-    /// Overrides button down.
-    fn button_down(&mut self, _: i32, button: Button) {
-        match button {
-            Button::Mouse(mouse_button) => {
-                if mouse_button == MouseButton::Left {
-                    if self.selected_state {
-                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
-                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
-                    } else {
-                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
-                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
-                    }
-                }
-            }
-            _ => (),
-        }
-    }
-
-    /// Overrides button up inside, triggering an `on_selected` callback.
-    fn button_up_inside(&mut self, _: i32, button: Button) {
-        match button {
-            Button::Mouse(mouse_button) => {
-                if mouse_button == MouseButton::Left {
-                    self.call_on_selected();
-
-                    if self.selected_state {
-                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
-                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
-                    } else {
-                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
-                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
-                    }
-                }
-            }
-            _ => (),
-        }
-    }
-
-    /// Overrides button up outside.
-    fn button_up_outside(&mut self, _: i32, button: Button) {
-        match button {
-            Button::Mouse(mouse_button) => {
-                if mouse_button == MouseButton::Left {
-                    if self.selected_state {
-                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
-                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
-                    } else {
-                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
-                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
-                    }
-                }
-            }
-            _ => (),
-        }
-    }
+//    /// Overrides button down.
+//    fn button_down(&mut self, _: i32, button: Button) {
+//        match button {
+//            Button::Mouse(mouse_button) => {
+//                if mouse_button == MouseButton::Left {
+//                    if self.selected_state {
+//                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
+//                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
+//                    } else {
+//                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
+//                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
+//                    }
+//                }
+//            }
+//            _ => (),
+//        }
+//    }
+//
+//    /// Overrides button up inside, triggering an `on_selected` callback.
+//    fn button_up_inside(&mut self, _: i32, button: Button) {
+//        match button {
+//            Button::Mouse(mouse_button) => {
+//                if mouse_button == MouseButton::Left {
+//                    self.call_on_selected();
+//
+//                    if self.selected_state {
+//                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
+//                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
+//                    } else {
+//                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
+//                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
+//                    }
+//                }
+//            }
+//            _ => (),
+//        }
+//    }
+//
+//    /// Overrides button up outside.
+//    fn button_up_outside(&mut self, _: i32, button: Button) {
+//        match button {
+//            Button::Mouse(mouse_button) => {
+//                if mouse_button == MouseButton::Left {
+//                    if self.selected_state {
+//                        self.base_widget.set_color([0.0, 0.0, 0.0, 1.0]);
+//                        self.text_widget.set_text_color([1.0, 1.0, 1.0, 1.0]);
+//                    } else {
+//                        self.base_widget.set_color([1.0, 1.0, 1.0, 1.0]);
+//                        self.text_widget.set_text_color([0.0, 0.0, 0.0, 1.0]);
+//                    }
+//                }
+//            }
+//            _ => (),
+//        }
+//    }
 
     /// Draws the contents of the widget in this order:
     ///